requirements.txt
streamlit>=1.28.0
lxml>=5.1.0
deep-translator>=1.11.0
python-docx>=1.1.0
Pillow>=10.0.0
requests>=2.31.0
